refactored code for all directories
